Moved configureRootProjectProperties to occur during evaluation and before configureExtensions.
This change was motivated by wanting to be able to refer to root-project properties during configureExtensions.
configureExtensions currently determines which tests to add, and more-specifically, builds-out the test-container
matrix.  It is useful to have access to root project properties here so that the matrix can be filtered if desired.

This change seems especially appropriate considering that configureRootProjectProperties is currently a NOOP after
the first sub-project is configured with the plugin, which sets the root project properties for the first time. That
is how this change was noticed - the first subproject was the only one which did not have access to the set of root
project properties.