Set initial log level (fixes #150)
Prior to this patch, Gretty uses Logbacks default log level
of DEBUG until the server start command had been received,
and Gretty reconfigures Logback according to the server
parameters.

This has caused a situation in which log messages prior to
server start are useful (command line arguments, network
setup, commands received other than 'server start'), but
were not properly thresholded.

This patch establishes a log level for all log messages
prior to receiving the server start command. To this end,
the runner accepts a "--debug" command line flag to turn
on debug logging, and sets INFO otherwise.
The Gretty Gradle process invoking the runner turns on the
debug flag only if its process itself was invoked with
"--debug", as in `./gradlew --debug :helloGretty:appRun`.

This type of setup is repeated in the default logging
configuration in `grettyRunnerLogback.groovy`.
If the server start parameters contain a different Logback
configuration file, it takes precedence over the debug
flag.