Fix behavior of maxPostSize in Tomcat 10 (#144)
Connector#setProperty is a bit of a lie. It does not set properties
on the Connector. Rather, the method set properties on the protocol
handler which is part of the connector.
Thus, the previous implementation tried to set a connector property
using setProperty, which was never successful.
Instead we directly set the property of the connector. To achieve
the original intent ("unlimited post size") the value must be
below zero. See:
http://tomcat.apache.org/tomcat-10.0-doc/config/http.html#Common_Attributes