Set backwards-compatible default values again
Commit dfa4142340c798e41d603dfb250d7f2c61077279
temporarily removed Jetty, because it was not ready
for Jakarta yet. As a result, some default values
could not be kept.
Because we now have Jetty 11 integrated, the default
servlet container can again become Jetty. This
change makes Gretty 4 a bit more backwards compatible,
because the default values did not change, and
configurations relying on these defaults continue
to work as expected.